EmpowerNormalizer Updates
- Made the EmpowerNormalizer Global
- Added Slicing Winds and Azerite Surge (lmao no one uses that)
