+++ src/analysis/retail/evoker/shared/constants.ts
@@ -21,22 +21,6 @@ export const TEMORAL_BURST_CDR_MODIFIER_PER_STACK = 0.01;
 export const EMPOWER_BASE_GCD = 1500;
 export const EMPOWER_CANCELED_GCD = 750;
 export const EMPOWER_MINIMUM_GCD = 500;
-export const EMPOWERS = [
-  // Shared
-  SPELLS.FIRE_BREATH.id,
-  SPELLS.FIRE_BREATH_FONT.id,
-  // Devastation
-  SPELLS.ETERNITY_SURGE.id,
-  SPELLS.ETERNITY_SURGE_FONT.id,
-  // Augmentation
-  SPELLS.UPHEAVAL.id,
-  SPELLS.UPHEAVAL_FONT.id,
-  // Preservation
-  // TALENTS.SPIRITBLOOM_TALENT.id,
-  SPELLS.SPIRITBLOOM_FONT.id,
-  TALENTS.DREAM_BREATH_TALENT.id,
-  SPELLS.DREAM_BREATH_FONT.id,
-];
 
 export const EB_BUFF_IDS = [
   SPELLS.ESSENCE_BURST_BUFF.id,
