+++ src/parser/shared/modules/GlobalCooldown.tsx
@@ -21,10 +21,7 @@ import { wclGameVersionToBranch } from 'game/VERSIONS';
 import GameBranch from 'game/GameBranch';
 import { BadColor, GoodColor, OkColor } from 'interface/guide';
 import SpellLink from 'interface/SpellLink';
-import {
-  EMPOWER_CANCEL,
-  EMPOWER_END,
-} from 'analysis/retail/evoker/shared/modules/normalizers/EmpowerNormalizer';
+import { EMPOWER_CANCEL, EMPOWER_END } from 'parser/shared/normalizers/EmpowerNormalizer';
 const INVALID_GCD_CONFIG_LAG_MARGIN = 150; // not sure what this is based around, but <150 seems to catch most false positives
 const MIN_GCD = 750; // Minimum GCD for most abilities is 750ms.
 const MIN_GCD_CLASSIC = 1000; // Minimum regular GCD was 1s until Legion
@@ -53,8 +50,8 @@ class GlobalCooldown extends Analyzer {
     this.addEventListener(Events.cast.by(SELECTED_PLAYER), this.onCast);
     this.addEventListener(Events.BeginChannel.by(SELECTED_PLAYER), this.onBeginChannel);
     this.addEventListener(Events.GlobalCooldown.to(SELECTED_PLAYER), this.onGlobalcooldown);
-    this.addEventListener(Events.empowerEnd.by(SELECTED_PLAYER), this.onEmpowerEnd);
-    this.addEventListener(Events.empowerCancel.by(SELECTED_PLAYER), this.onEmpowerEnd);
+    this.addEventListener(Events.empowerEnd.by(SELECTED_PLAYER), this.onEmpowerTrigger);
+    this.addEventListener(Events.empowerCancel.by(SELECTED_PLAYER), this.onEmpowerTrigger);
 
     if (wclGameVersionToBranch(options.owner.report.gameVersion) === GameBranch.Classic) {
       this.minDuration = MIN_GCD_CLASSIC;
@@ -79,6 +76,7 @@ class GlobalCooldown extends Analyzer {
     return Boolean(this.getGlobalCooldownDuration(spellId));
   }
 
+  /** Returns true if this ability has the empower specific events linked, false if not */
   isEmpowerSpell(event: CastEvent | BeginChannelEvent): boolean {
     // LinkedEvents for Channels are nested deeper
     if (event.type === EventType.Cast) {
@@ -156,7 +154,9 @@ class GlobalCooldown extends Analyzer {
     event.globalCooldown = this.triggerGlobalCooldown(event);
   }
 
-  onEmpowerEnd(event: EmpowerEndEvent | EmpowerCancelEvent) {
+  /** Empower global cooldown handling for empower end and empower cancel events.
+   * Currently only evoker empower spells have these events linked / added*/
+  onEmpowerTrigger(event: EmpowerEndEvent | EmpowerCancelEvent) {
     event.globalCooldown = this.triggerGlobalCooldown(event);
   }
